ai-research-assistant/requirements.txt
streamlit
transformers
torch
beautifulsoup4
requests
python-dotenv
fpdf

